ml-service/requirements.txt
absl-py==2.4.0
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1769745019/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.133.0
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
idna==3.11
importlib_metadata==8.7.1
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1770566174/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
msgpack==1.1.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numba==0.63.1
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.1.0
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1770676243/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==12.1.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1771048953/work
pooch==1.9.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==3.19.6
psutil @ file:///D:/bld/bld/rattler-build_psutil_1769678167/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
Pympler==1.1
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-multipart==0.0.22
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.15.3
seaborn==0.13.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.52.1
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
tornado @ file:///D:/bld/tornado_1765836387919/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
validators==0.35.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1770634525285/work
Werkzeug==3.1.5
wrapt==2.1.1
zipp==3.23.0
